Refactor classes into modules
